Add an image: Attach the inspector to the article instead of the toolbar (desktop)
Be default, ToolbarDialog has the toolbar as its ancestor. On desktop, when the toolbar is animating in, the containing block for the fixed positioned inspector is the toolbar rather than the viewport since the toolbar has a transform property set. This causes the inspector to animate from the top of the page rather than from the bottom when the inspector appears before the toolbar is fully shown.

Bug: T300586
Change-Id: Id061b64d216d61a10f88e13e54a0b930a1b84bcd