Make WikiPageConfigLoader support loading a mentor list
This consists of writing and registering a validation logic
for the mentor list, as WikiPageConfigLoader only supports pages
that can be validated by a config validator.

There will be only one mentor list, stored on-wiki as
MediaWiki:GrowthMentors.json (can be configured through
GEStructuredMentorList). Whether a mentor is assigned
automatically or not will be controlled by a field called
`automaticallyAssigned`.

Each top-level key is an array<int,array>, where the int
is the mentor's user ID and array is an object describing
the mentor (including more detailed information about the mentor,
such as, their message to the newcomer).

I decided to use user IDs rather than usernames, because user IDs
are guaranteed to be stable, even when the user gets renamed, while
usernames can change at any time (without an easy way to update
the mentor list back).

The array (describing the mentor) can contain the following fields:
   * message: contains the mentor's introductory message (or null if the
     default one is to be used)
   * weight: number controlling mentor's defined weight (replaces
     MentorWeightManager)
   * automaticallyAssigned: a boolean controlling whether the mentor
     is automatically assigned to any newcomers.

Bug: T264343
Change-Id: Iec28a0bd513674e790e25666e09b59cf9a5cb6fd