Tag newcomer task edits via VisualEditor or post-edit dialog
What:

* Replace the TrackerFactory system for tracking newcomer task edits
with an implementation using VisualEditor and a REST endpoint for non-VE
newcomer edits.

Why:

When a user makes an edit to an article via Special:Homepage Suggested
Edits, we tag it with "newcomer task" and other change tags that are
more specific ("add links", "add images"). The way we track this is a
hacky, best-case approximation: when a user clicks on an article, we add
an entry to a per-user cache with a TTL of a week. Then, in the
RecentChanges save hook, we look to see if the recent change object has
a match in that user's cache, and if so we assume the user must have
made the edit via Suggested Edits; the rationale being that the user
would be very unlikely to have otherwise edited that page through some
other workflow.

This has mostly worked, but there are weird edge cases where
occasionally edits that don't appear to be e.g. "add image" edits are
nonetheless tagged with "add image". We can be more precise about how we
are tagging suggested edit edits, and that's what this patch is
implementing.

How:

* Create edit tags for each template-based task type (e.g.
  "newcomer task copyedit"). Those are applied in addition to the
  generic "newcomer task" tag.
* For VisualEditor edits, use VE's PostSave hook to tag the edits with
  "newcomer task" and task type specific change tags
* For non-VisualEditor edits, use the post-edit dialog loading process
  to send an an HTTP request to an endpoint and handle edit tags in
  the handler for that endpoint.

Misc:

* Rework Selenium tests to assert success for tagging edits in this way
* Remove two homepage Selenium tests that are not providing value

Bug: T297004
Change-Id: I289c3a5ea759c97775eca1b31929d1c7f42361c2