Newcomer tasks: add topic match mode parameter for filtering
What:
* modify getQueries in SearchStrategy to accept an extra
parameter $topicsFilterMode (one of 'AND', 'OR'). Using
'AND' will create a single query per task type with
all topics. Using 'OR' or uninformed mode will fallback to
the existing strategy where we create n_topics * n_task-types
queries
* Add url query parameter 'topicsmode' to ApiQueryGrowthTasks
which accepts 'AND' and 'OR' values. The search strategy will
use according mode for topics

Misc:
* we might run again into query length limitations (see T242560)
if a user selects many topics (All ores topics length is ~800 chars
and query limits ~300).

Bug: T301825
Change-Id: I44ff56aacea8bf6e6bb2fa0ff9181b9ce29f27dd