SpecialEditGrowthConfig: Do not require GEInfoboxTemplates to be templates
This also ensures Special:EditGrowthConfig stores full page titles whenever
possible. It intentionally keeps the default namespace
in TemplateCollectionFeature, to not break any of the current configs

I tried to add the Template: prefix on-wiki before a patch lands at the wikis,
but it turns out that:

    a) Without this change, I can't add page that's in a non-template NS
    b) After doing a null edit via Special:EditGrowthConfig, the Template:
       prefix is removed (which makes doing it before code changes harder,
       as any change by a community would inevitably revert that).

To avoid that, I decided to first ensure full page names are stored whenever
possible, then to change the on-wiki config to no longer rely on NS_TEMPLATE
being the default and then we can remove the default namespace
from TemplateCollectionFeature.

Bug: T298390
Change-Id: I127eecc5a9a147762f53d60e6ffdbf3be129f4e9