WikiPageConfigWriter: Remove the JsonContent TODO
As of 231073c, JsonContent is newable. As of f5623b6,
JsonContent is documented as a proper solution
for "JSON text content that can be viewed and edit [sic]
directly by users". This is how community configuration
works – while we provide tooling for communities to edit
their configuration easily (via Special:EditGrowthConfig),
they can always view and edit the actual JSON easily.

Change-Id: I11f2540a70f6ae8a131af9f8b8fb37d9c8e7ee6c