RecentChanges: Fix newcomer task tags display
In I289c3a5ea759c97775eca1b31929d1c7f42361c2 we re-worked how change
tags are applied to edits made via Special:Homepage's suggested edits
module, a.k.a. "newcomer tasks".

That worked, except that thanks to T24509, the change tags don't appear
in RecentChanges.

It is possible to set the edit tags by passing the 'tags'
parameter in the edit. Sounds good in theory. Except that if you do this
with software defined tags, ApiEditPage will complain that you don't have
permission. Implementing the onChangeTagsAllowedAdd hook could do this
for us, but since we have to inspect the request object, we may as well
re-implement RecentChange_save hook to make this more straightforward.

Misc other changes:

* Update the Selenium tests to make a follow-up edit to the copyedit
article, and to click on the card in the post-edit dialog to edit
another article afterwards. That's needed in order to verify the REST
handler that starts a new session.

Bug: T304747
Change-Id: I494d72a42d9103c28c4d44077cfe0f1269fc7b00