Mentee overview: Read mentee overview filters from the presets option if available
Previously, persisted filters were saved in their own user option,
called `growthexperiments-mentee-overview-filters`. After this change,
the Mentor dashboard will first attempt to read them from
`growthexperiments-mentee-overview-presets`. If that is not possible,
`growthexperiments-mentee-overview-filters` will be consulted.

The next step is to fully get rid of `growthexperiments-mentee-overview-filters`
by a maintenance script that will migrate it
to `growthexperiments-mentee-overview-presets`.

Bug: T304057
Change-Id: I4c5e58f612f1462b1d56547e299b0f99f140b81e