Account creation: allow campaign configuration via GECampaigns option
* Use skipWelcomeSurvey index with bool value inside campaigns config
for a specific campaign to allow the accounts created within to skip the
welcome survey.
* Use qualityGateIdsToSkip inside the campaign configuration. ie:
  'qualityGateIdsToSkip' => [
    'image-recommendation' => [ 'dailyLimit' ]
  ]
Set inside a specific campaign to allow accounts created within
to skip the daily limit for the given task type.
* Use messageKey index with string value
* Remove GLAM FIXME comments

Migrating old campaigns:
* The existing marketing video campaign (social-latam-2022-A) should
be turned into a campaign config entry, with:
   'messageKey' => 'marketingvideocampaign',
   'pattern' => '/^social-latam-2022-A$/',

* Any other prior campaigns willing to skip the welcome survey need
to add the following config options:
  'messageKey' => 'key-to-use-for-copy-text',
  'pattern' => '/^campaign-pattern-used-in-url$/',
  'skipWelcomeSurvey' => true,

Note that the default for skipping the welcome survey has changed within
this patch to show the welcome survey.

Bug: T305443
Depends-On: I7f8401533c02f7c4c68fcc2ef68480649e4bfb5d
Change-Id: Iea95a7e56415657dddef4a25f6159527d18aeb03