Suggested edits: move non-UI logic out of ext.growthExperiments.Homepage.SuggestedEdits
* Create ext.growthExperiments.DataStore module for state management
This can later be written as a Vue store when the migration happens. Once this patch is merged, other modules that currently have duplicated logic for fetching tasks (ext.growthExperiments.Homepage.mobile etc) can be refactored to use this module.

* Create stores for managing suggested edits tasks and filters
Information about the task queue, which task the user is viewing, which filters are selected are stored separately in the stores. ext.growthExperiments.Homepage.SuggestedEdits doesn't directly store these states. Instead it reads and writes data from the stores and updates the UIs when the data changes.

Bug: T303985
Change-Id: Iae154c7911da5231cf1958be385e8aa578c1da19