Video landing page: Fix broken style & HTML structure
Remove a display:none that should not have been there, but fortunately
didn't work, because the paragraphs of the body get reparented by the
browser because .mw-ge-donorsignup-body is a <p> tag and cannot contain
flow content. Fix that by using a <div> instead.

This will result in some spacing changes but we did not have exact specs
for that and the original spacing was also somewhat accidental, the
patch doesn't try to preseve it. It will also result in a slight text
color change on Minerva as that CSS rule was ineffectual until now.

Bug: T303785
Change-Id: I34cba569e99bdfc353e7853b8e0108dbecacbdf2