MenteeOverviewPresets: Handle invalid JSON
Even though `growthexperiments-mentee-overview-presets` is
not directly settable from the UI, it's still an user property,
which means users can control what is stored there via
the action=options API. As such, no assumptions about
the option's content should be made.

If JSON.parse throws, return an empty set of presets,
which should result in a working mentee overview module
(the homepage does not expect any presets to be set).

Change-Id: Idd580dd48cac78ac576f746f58815672233b405c