Add ApiManageMentorList
This will be used in the Mentor dashboard's client-side dialogs
to let mentors to manage their entry in the structured mentor list.

The newly-introduced `enrollasmentor` right will be used to control
whether mentors can sign up to the list. We will be able to
make use of MW's built-in groups like autoconfirmed, use MW's
autopromotion mechanism or write our custom logic as part
of an UserGetRights hook implementation.

Bug: T264343
Change-Id: Iefbb309b45a3ced5a622231803ed4fd47d39f677