GrowthTasksApi: Set topics match mode in API params separately from topics
If the user's topic list is empty and the user had set match mode to
AND, what happens is that the cached task set on the server side has the
"AND" match mode, but the client-side request uses "OR", so the cached
task set is never used.

To fix that, topics match mode is sent with the API params independently
of whether the user currently has any topics set. This scenario can
occur if the user selects topics + "AND" match mode and then deselects
topics. (Or if the user was in a campaign with custom topics and those
topics are then removed.)

Bug: T308312
Change-Id: I59517546e54ac4a4d17f4428db09f840f1a56b82