SuggestedEdits: Export task queue from server-side
* Export the task set from the server-side. Immediately refresh the task
set (via a DeferredUpdate) so that the user has a fresh, shuffled set of
tasks on subsequent page reload.
* Don't make an additional API request when suggested edits loads

Benefits:

* One less API request
* Improved performance for users on high latency connection as task
queue is interactive without needing to wait for an API request
to complete. It's unclear how often it would be the case that JS
would load successfully while the API request to fetch tasks would fail,
but at least in scenarios where the device has cached JS, the user has a
faster time to being able to interact with the feature

Drawbacks:

* Higher likelihood that the user gets an invalidated task from the
unstructured pool of tasks (e.g. the "copyedit" template for an article
removed). Unsure how often that happens in practice. It could be
mitigated by 1) bringing back the TemplateFilter (it had its issues,
see T267216 and T264029 but perhaps it is not a problem operating on
sets of 20 articles at a time?) or 2) implementing a
"No suggestions found" UX for unstructured tasks when a user clicks on
e.g. a copyedit task and the article doesn't have any of the copyedit
templates present.

Bug: T308542
Change-Id: Iaff0218ecddfd5b825a30ea2b9d56bfc055eedec