Make NewcomerTaskCompleteHandlerTest use HandlerTestTrait
The trait can be used to avoid reinventing the wheel, that this test was
kinda doing. Also, the test was calling Handler::init() directly, but
that's an internal method and calling it makes it more difficult to
change its signature (see Ic6bd48b400ecd839ef99b518ef955781470cd05c).

Also avoid passing RequestContext around, that class is a kitchen sink
and pretty much the nemesis of DI. Certainly not something that you want
to inject.

Change-Id: Ic7c1b19b86e8a151e2d42aaec00ef0e89db77f08