Add rescore method for sorting by underlinkedness
Adds two new link recommendation task type properties,
underlinkedWeight (default: 0) and underlinkedMinLength.
(Currently not exposed to Special:EditGrowthConfig.)

Defines a new rescore profile, growth_underlinked, which
uses a scoring function to sort underlinked articles
first. The scoring formula is roughly

(1 - (links in article / words in article))^4

which is then averaged with the given weight with a
purely random score. For articles shorter than
underlinkedMinLength bytes, the non-random part is 0.

Updates the search suggester, query and strategy
classes to be able to specify a rescore profile.
When the query is for link recommendation tasks,
isn't limited to a specific pageid set (ie. it is
not a cache refresh) and underlinkedWeight is not 0,
new rescore profile is used.

Bug: T301096
Change-Id: I51f785fa5060a18c76c4fcf20f402dd7c404f57f
Depends-On: Id0b57312715069d13325f041872b74b429bec3c0
Depends-On: I33746a0dfa2448dc557dba623461ac4b4162244f