SubmissionHandler: Allow null for baseRev ID
In certain esoteric cases, it's possible that VisualEditor's
'oldid' (the base revision ID) will be null when the attempt is made to
obtain it from the revision store.

To account for that, we can loosen the signature for handle() and
validate() to allow for null; the only handler that needs its business
logic adjusted for this scenario is link recommendation.

Bug: T309120
Change-Id: Id1614cceab017b58450f38bd35638763a8c6e776