WelcomeSurvey/VariantHooks: Change hook used for redirection
When a user signs up for an account, in most scenarios,
GrowthExperiments should redirect the user to Special:WelcomeSurvey.

Prior to this patch, we did that by issuing a redirect via the
RequestContext OutputPage in the BeforeWelcomeCreation hook. This
likely conflicted with other extensions trying to redirect elsewhere
(specifically CentralAuth, for central login).

It would be more correct to use one of CentralAuthPostLoginRedirect,
when CentralAuth is enabled, or PostLoginRedirect when it is not.

To test this, try enabling and disabling CentralAuth while creating an
account from an article page. The behavior should be identical to what
currently happens in the main branch.

Similarly, test creating an account using a campaign configuration with
skipWelcomeSurvey enabled/disabled; the behvaior should be identical to
current production.

One change in this patch is that, when not in a ResourceLoader context,
we look up the Growth campaign from the user's preferences and fall back
to the query parameter in the request. That's because the campaign query
parameter would sometimes get lost in the process of redirection to and
back from the login wiki.

Bug: T313064
Change-Id: Idc645aefcd84b9ae6ff9eed892dcbcb11402d6f6
(cherry picked from commit 8921c7584c4991bd8b2a4f29050070f111b1d22a)