Make EllipsisMenu a singleton
Previously, MobileOverlay and ext.growthExperiments.Homepage.Mentorship
used to have their own EllipsisMenu instance. This didn't work well
when we needed to use EllispsiMenu's features from outside (to
make the menu think it was clicked by the user, while the user
clicked a link that was outside of the menu), because
there were two instances of the same dialog (one for each instance
of EllipsisMenu).

This patch avoids that problem by making EllipsisMenu
a singleton.

Bug: T292627
Change-Id: I42654553b27d4817d2ad924854b9da068fa162a7