Move wrapping intro text to quotation marks into Mentorship class
Previously, MentorProvider wrapped mentor's intro text
with quotation marks. This is not ideal, since MentorProvider
is meant to provide information about mentors, not to format
them. This patch moves the logic to the Mentorship class.

This change is done mainly to make writing
migrateWikitextMentorList.php easier (T310446), as
the quotation marks are not supposed to appear
in the resulting JSON.

Change-Id: I1769b734c340696afa3884c6b89d97d70ed6cb1f