ApiManageMentorList: Empty message should mean default
The current API doesn't make it possible for mentors
to switch to their default message. Null (used by the Mentor
value object) is not usable, because it is interpreted
as "do not change". As it does not make sense to use
an empty custom message, let's use it to mean
"use the default mentor message".

Bug: T264343
Change-Id: I7086e9ed38fe623a69a9c85dab8be997f0547789