Rename SpecialQuitMentorship to SpecialQuitMentorshipWikitext
Instead of adding a bunch of ifs to the special page's class,
let's add two new classes:

    * SpecialQuitMentorshipWikitext (when the old mentor list
      is in use)
    * SpecialQuitMentorshipStructured (when the new mentor list
      is in use)

User traffic to Special:QuitMentorship will be routed to
SpecialQuitMentorshipWikitext or SpecialQuitMentorshipStructured
depending on the current mentor provider.

That way, any potential wiki-specific workflows that depend
on the special page name won't break, even after we switch
the mentor provider to structured.

Bug: T311747
Change-Id: Iad6df32a5604fc23c486d2949e95448fdcf35aaf