Add SpecialQuitMentorshipStructured
Intended behavior of SpecialQuitMentorship is very different,
depending on whether structured mentor list is enabled or not.
To account for that, let's temporarily introduce two classes
for the same special page. Once we're fully migrated
to the structured mentor list, we will delete
SpecialQuitMentorshipWikitext and rename
SpecialQuitMentorshipStructured back to SpecialQuitMentorship.

Bug: T311747
Change-Id: I7be864205ccc4fd8146163a96d78a908ae503b3d