Mentee overview(Vue): remove custom menu item template
The problem was in the the custom menu item template. When
trying to split the results in chunks to highlight the matched
fragment there was no proper case sensitivity handling.

Avoid higlighting the matched part for now since it should
be supported by Codex directly and this is not implemented
neither in the non/vue version of the mentee overview.

Bug: T311992
Change-Id: Ia2683fa3a11a0dea3dcf5add47359a2e33246eab