QuitMentorship: Do not check if a mentor has mentees with structured list
The only way for mentors to quit when a structured mentor
list is enabled is to use Special:QuitMentorship. Currently,
mentors are unable to use this page if they do not have any mentees.
This is not correct, because this leaves them unable to sign up
unless they claim someone (or are assigned someone), which is absurd.

This patch changes this behavior, and lets all current mentors
to use Special:QuitMentorship.

Bug: T312577
Change-Id: Ice7b2162cc78539ec7fa4cf221eb84f4335ee30b