Update to accomodate future changes to JsonCodec
JsonCodec will handle recursion itself in the future, so you can directly
include JsonSerializable objects as values in the array returned by
::toJsonArray(), and those objects will already be unserialized values
in the array given to ::newFromJsonArray().

This patch makes the code in GrowthExperiments agnostic about whether
unserialization occurs before or after ::newFromJsonArray() occurs,
in order to handle both the old and proposed new behavior of JsonCodec.

Bug: T313452
Change-Id: I3413609251f056893d3921df23698aeed40754ed