Make use of PHPUnit shortcuts like create(NoOp)Mock
… and inline the resulting 1-line methods.

I used either createMock() or createNoOpMock() as a replacement
depending on the complexity of the class and what it represents.
Larger "god classes" like WebRequest that can access global state
are good candidates for createNoOpMock to make sure no unwanted
code is called, even if this is not strictly necessary. Small
interfaces or value objects don't need this.

Change-Id: I34ba3b14b0aa8b80cd955020c2b71355efbdde30