MentorDashboard.Vue: allow setting empty values for edit count filters
Prior to this change it is not possible to unset the values
for the edit count minimum and maximum in the mentee filters form.
Allow the user to remove these values and use '' empty string
to inform the mentees store the filter needs to be removed
in the next API request.

Also replace required:true by default:undefined for the
caption text prop in data-table component to avoid
useless warnings, the prop should be optional as per
requirements from T311393. Follow-up of Id0598cfd98a642e2cfff771556d8df6b99a3cd4a

Bug: T313263
Change-Id: Ie35110c7b36882c847490cc28599f27077ee4763