Post edit dialog: update tasks fetched accumulator based on API response
Prior to this change we were wrongly assuming the taskQueue has been
already set when we call NewcomerTasksStore.fetchTasks(). That assumption
is correct for the homepage but not for the post edit dialog which
appears on the article page and no pre-loaded tasks are used.

Use always the length of the received collection of tasks in fetchTasks()
call so the counter is reset on each call. This covers scenario (2) described
in T312686.

Bug: T312686
Change-Id: Ia1e65b681cdc0273ed414592bd71964d7d86868c