Delay showing the welcome survey if the user was editing
Instead of disrupting an editing workflow by showing the
welcome survey, remember that it should be shown, return
to the edit page and show a notice to the user after saving
the edit.

To do that, we need to 1) detect that we are in a pageview
right after a save, 2) we need to remember that the user
registered during the edit. There aren't elegant solutions
for either. For the first, we rely on core's post-edit
mechanism (which wasn't really meant for server-side
integration, but it does require server-side integration
so the client-side hook handler gets loaded). For the
second, there are a couple options:
* Use a cookie, which is easy to handle but is not tied
  to the browser window so there can be all kinds of
  unexpected effects.
* Use sessionStorage, but that cannot be detected on the
  server-side.
* Use an URL query flag, but that's hard to maintain during
  editing.

We go with a combination of the sessionStorage and cookie
options, to track the browser window but minimize how often
we need to add an extra module on the server side.

Bug: T310320
Change-Id: I73181766c2e9484725575c8784b22aeefce8f0fd
Depends-On: I8864803529b5654910bd671b2286bd6ea76f977e