MentorTools: Ensure weight setter displays correct weight
It looks that findItemFromData() uses a datatype-sensitive
comparsion, which means we need to ensure weight is a number,
not a string.

This solution uses a ternary condition, based on
isNaN( parseInt( weight) )'s result. This is needed
to avoid issues with the "none" weight (appears
with structured mentor list).

If mw.user.options.get guaranteed to return an integer
when possible, we'd be able to use that function (that
is not the case, unfortunately).

Bug: T314050
Change-Id: I8f592946c874edb8c886d315e51054d332479357
(cherry picked from commit 1ed0edbc2c73d339b094f41f8c62825e0dc9e9f5)