MentorWeightManager: Throw when constructed with wrong provider
MentorWeightManager provides incorrect (out of date)
information when structured mentor provider is in use.
Throw a LogicException in that case, as that's easier to notice
in logging than incorrect information being used.

Parent patches should ensure the only class
depending on MentorWeightManager is WikitextMentorProvider,
which is only constructed when the wikitext provider
is in use.

Bug: T314362
Change-Id: Ie356c8aa29ca50783a42ec9e0edbda1740c4d938