Mentee overview(vue): use number input for the edit count filters
In the non-vue version of the mentee overview the edit count
uses ooui's NumberInputWidget which has custom validation
and uses type=number for the input. This prevents the user
from input non-numeric characters.

For negative integers the input still allows to input
'-' minus sign but shows the input in an invalid state.

Bug: T300532
Bug: T313263
Change-Id: I1406e67b694538c3c88d149617c56e2c20f27f93