Move WEIGHT_* constants to IMentorWeights
Previously, only MentorWeightManager was responsible
for determining which mentors have which weight. With
the structured mentor list, this responsibility is now
shifted:

    * Wikitext mentor list: MentorWeightManager, as before
    * Structured mentor list: StructuredMentorProvider

This is because weight is no longer stored in user preferences
when structured mentor list is in use. Instead, it lives in
the MediaWiki:GrowthMentors.json JSON page, as part of
the declaration of a mentor.

When structured mentor list will be fully deployed, we
will delete MentorWeightManager, but the constants will
still be needed. Move them to a separate interface.

Bug: T314362
Change-Id: I213feff6b55a577dec68bc39893d3b31735194b4