Show no suggestion dialog when page is protected
Handles edge cases such as the article getting after the user loads
the task queue but before clicking on a card. Previously, the edit
interface loaded and the user got an error when trying to save.

For non-structured tasks we have no equivalent of the error dialog,
but also don't automatically open the editor, so those will just
drop the user on an uneditable page in view mode, which is not
great but less problematic.

Since this is a feature aimed at new users, no effort is made to
handle differently the case when the page is protected but the
user does have the rights to edit it.

Change-Id: I98d8865aed164afa7ab235c3d1a86144957d9c9b