Newcomer tasks: dont emit update event until count is set
Prior to this change, in NewcomerTasksStore.fetchTasks, this.taskCount
was set in two different places (lines 313 and 316). The store emits event
when calling this.setTaskQueue so components listening to the queue changes
were not getting the last calculated taskCount for the scenario where
the API gets less results than requested. This covers scenario
(1) in T312686.

Tests in NewcomerTasksStore.test.js should be reviewed.

Bug: T312686
Change-Id: I88ca960aaf45da41a63cb7722d9caddcb4b33638