MenteeOverviewApi: Respect uselang parameter
Previously, calls to the REST API endpoint
always used the language user has in their
preferences. This works fine in most cases,
but not when uselang is passed to Special:MentorDashboard.

For simplicity, let's pass uselang to REST
always, not only when current language is
different from the one in preferences,
as that's harder to detect on the client
side.

Bug: T314849
Change-Id: I6b1e702a4c85c59873f0b582564027fb2b1840e9