MenteeOverview: Do not use mw: for linking to MW.org
The Popups extension is unable to determine the link
is external, because it is not possible to parse
links like [[:mw:Growth]] from the JS due to lack
of interwiki table on the client side.

Instead of using [[:mw:Growth]] type of links, use
[https://mediawiki.org/wiki/Growth Growth] links. This
makes mediawiki.jqueryMsg to add class="external",
which in turns stops Popups extension from triggering
the popup (see extensions/Popups:src/index.js:32, where
it excludes selector `.external`).

Bug: T312248
Change-Id: Ia5f205f85bd910f70c0fc7d34da57e95b49cfe29