WikiPageConfigLoader: Stop using local cluster cache
pcTTL is not used, because it doesn't appear
to be invalidatable. This behavior was breaking
WikiPageConfigLoaderTest.

We probably can cache errors (and hope
WikiPageConfigWriter and/or ConfigHooks::onPageSaveComplete
will invalidate cached errors when needed), but that
would be a functional change, so I decided to
not do that as part of this refactoring commit.

Bug: T298328
Change-Id: I78687a2e996175f6d0894944ffb0882bca6c41db