Add maintenance script to refresh user impact data
Select users who have registered recently and/or edited
recently, and update their user impact record.

Add a 'generatedAt' field to UserImpact to keep track of
age, and allow skipping users with recent user impact
records when running the script. Also add a versioning
scheme to UserImpact JSON export/import so future breaking
changes can be handled gracefully.

Query plans:
* https://phabricator.wikimedia.org/P34890
* https://phabricator.wikimedia.org/P34891
* https://phabricator.wikimedia.org/P34892

As is, this is useful for development / on small wikis.
It is disfunctional on large wikis (e.g. on stat1006 enwiki
takes 1 min / batch) so the queries need to be improved.

Bug: T313395
Change-Id: I6a61fb871e4a1738ea0fe3237c26882bbf45aaf5