MentorTools: Make the link to QuitMentorship easier to find
This replaces the info icon in the Mentor tools module
with an ellipsis menu widget, to make it more obvious
that it empowers mentors to take a specific action.

We decided against putting it into the "Status" dropdown,
so we don't accidentally tell mentors to quit, when a temporary
mentoring-vacation is what they were originally looking for.

Bug: T312122
Change-Id: I2f10598e02b5059c8016c70cb18c9fc768f64b42