RefreshUserImpactData: Fix queries that filesort
Follows up I6a61fb871e4a1738ea0fe3237c26882bbf45aaf5.

Only GROUP BY on actor_user so that clause does not span
multiple indexes. This introduces fields which are present
in WHERE but not GROUP BY, not sure if that would work in
strict mode (the fields are all from the actor table which
is groupped on its primary key so in theory a sufficiently
clever query planner should figure it out), but if that
becomes a problem it's easily solved by only fetching the
actor_id and getting the other data in a second query.

Also get rid of ActorMigration, which is apparently not doing
anything useful anymore.

Query plans:
* https://phabricator.wikimedia.org/P34943
* https://phabricator.wikimedia.org/P34944
* https://phabricator.wikimedia.org/P34939

The two --registeredWithin queries are still not great when
there are lots of users as they need to scan the whole actor
table. --editedWithin seems very fast, somewhat surprisingly
given that it would either have to scan all recent revisions
(on enwiki something like 2M per day) or scan all actors
(the query plan indicates that it does the latter) but on
stat1006/enwiki it only takes tens of milliseconds.

Bug: T313395
Change-Id: Ic428f67c771ffe4bb8caf1d4912bef9673e1c0b6