User impact: create articles edited trend chart
* Add a minified version of d3.js as a foreign
resource used in ext.growthExperiments.d3 module. The
module will only be registered in onResourceLoaderRegisterModules
if GENewImpactD3Enabled is set to true. The client checks the
same GENewImpactD3Enabled config and try to load the module accordingly.
* Set a fixed width on the c-sparkline component since using
the parent or element width does not work for async components and
how we render the module inside the MobileFrontend overlay.
* Rename index.js to conventional init.js for application boot.

Bug: T310660
Change-Id: Id3aaf430f73aef7ce0de40bb493aea12c4fc8add