Improve logging for StructuredTaskPreEdit.showErrorDialogOnFailure
Don't assume the error has already been logged on the backend, as
that's not true on all code paths; instead allow an explicit flag
to be passed along to signal whether the error needs logging.

Also change callers to pass an Error instead of a string so the
stack trace is more informative, and mark the first parameter of
StructuredTaskPreEdit.showErrorDialogOnFailure optional as it is
actually used that way.

Change-Id: Ie65cad3ce5b85232b27b36b8fea33e7a66031b35