ReassignMentees: Pass the actual performer to ChangeMentor
With the structured mentor list (and T321382 in particular),
we want to allow admins to fully remove any mentor. This includes
reassigning all their mentees to someone else.

Prior to this patch, ReassignMentees claimed the mentor
is the performer. This is the case when a mentor retires,
but not for admin-initiated removals.

Add $performer to ReassignMentees, to properly
attribute the log entries. This cannot use IContextSource
(which is already injected into ReassignMentees), because
the context is replaced with a different one in ReassignMenteesJob.

(cherry-picked from 5db0092ecdb08b94757a01b4c051d7f9e9e54ac4)

Bug: T321382
Change-Id: Ibdcc93601457e50a08f2f2d4304b9e9ac78bcb9f