ReassignMentees: Pass the actual performer to ChangeMentor
With the structured mentor list (and T321382 in particular),
we want to allow admins to fully remove any mentor. This includes
reassigning all their mentees to someone else.

Prior to this patch, ReassignMentees claimed the mentor
is the performer. This is the case when a mentor retires,
but not for admin-initiated removals.

Add $performer to ReassignMentees, to properly
attribute the log entries. This cannot use IContextSource
(which is already injected into ReassignMentees), because
the context is replaced with a different one in ReassignMenteesJob.

Bug: T321382
Change-Id: I7e3884d8a6d4be2df56d07844cae03c8e71d18db