Localize default mentors' introduction correctly
When you visited Special:ManageMentors, the part of the default
mentor's introduction text that addresses you was localized based
on the mentor's gender, not yours.

Internally, the fallback to the mentor in case of null mentee was wrong.
Instead, use the empty string which corresponds to "{{GENDER:|you}}"
in wikitext.

The alternative would be to pretend the visitor is each mentors' mentee
by calling newMentorFromUserIdentity with their identity as the second
argument.

Bug: T322130
Change-Id: I8ac44b38dfc622941e05cb59bf5aae056bc6a9f5