UserImpactHandler: Load from ExpensiveUserImpact by default
Also remove the confusing "useLatest" flag. The two option are to load
inexpensive data (no page views) or expensive (page view data). So the
parameter option should be renamed, but thinking through it further, it
doesn't seem like putting the choice between expensive/inexpensive
behind a flag is that helpful.

Bug: T313393
Change-Id: I3c71775a0275a44c70e6a431fab569c35845f71b