User impact: add disabled and unactivated state
- Remove separated layouts since the style differences are more
appropiate to handle on the affected components and it was
replicating top level logic.
- Use User::getEditCount to determine if the user has done any
edits and display the unactivated state (0 edits so far).
- Pass the enabled and active flags to the NewImpact class
and as JSConfig vars to the client to display views accordingly.
- Require an id in TrendChart component. Since it's rendered
in overlay-summary and overaly modes each instance needs to
use a different id for the sparkline.

Bug: T321963
Change-Id: Id3c6064979b678ac24a1a3e32f55467f6ac72f3f