User impact: create separate layouts for each view of the module
To facilitate style and markup differences between the different
views of the module (desktop, mobile summary, mobile overlay) use a separate
layout for each to hold the ad-hoc styles and component selection.

Create an App wrapper to have a single place for top-level logic
like choosing the layout or displaying the unactive state.

Bug: T318542
Change-Id: I46951e808a96c4a285898e75cb859db953d175c9