UserImpactHandler: Re-calculate data on demand
Generate and cache the data on-demand:

- Continue to default to loading data from cache table
  - If page view data is stale (older than 48 hours), recompute data,
    and save it to the store via a job
- if no data is found, compute data and save to store via a job

To prevent abuse, apply rate limits for interacting with the
UserImpactHandler. The default is set to 10 requests per day.

With this change, we no longer need to run refreshUserImpactData.php,
though, this change is not incompatible with doing so.

We should continue to run deleteExpiredUserImpactData.php.

Misc:

- Remove unused time zone correction code from the page views tool URL
  generation method, it should be corrected in a different patch.

Bug: T313393
Change-Id: Ie769438581cd7b75a7159db91b016df59a8f1a15