MentorStore: Use $wgRCMaxAge instead of INACTIVITY_THRESHOLD
The "is active" flag will be used in the RC filters,
to avoid supplying more than 10k of user IDs (which
triggers warnings, see T293182).

It does not make sense to suply users
not editing for more than wgRCMaxAge
to recent changes, as RC will never
display such users, regardless of user-provided
filters.

Bug: T318457
Change-Id: I662aad2f007f1d2ad5dd2a0356970171c0425b07