NewImpact: Instrument app loading
Start a timer when App.vue begins executing, emit an event when the
NewImpact and NewImpactSummary components are mounted, and log the
difference between start / NewImpact mount time.

The display of the app contents is directly bound to the response
time of UserImpactHandler, but we will address that in T321675
which uses skeletons while data is loading. If/when we do that
task ,we might need to rework how this load time metric is
generated, because we'll want one metric for when the app renders
and a second one after data is rendered.

Bug: T310779
Change-Id: Iaae033eaaee599ce1d6a3131385f5f4fb65595b4