Ignore duplicates in mentorship related jobs
This sets ignoreDuplicates() to true in two jobs.

==== SetUserMentorDatabaseJob
Ocasionally, when the job queue is slow, multiple
instances of SetUserMentorDatabaseJob are enqueued.
Those jobs override each other, use write queries,
although only one result is preserved.

This patch turns the deduplication flag on. That will
not be sufficient to prevent T323374 and similar
issues, but it is the first step.

==== ReassignMenteesJob
ReassignMenteesJob exists to reassign mentees assigned
to former mentors. It is important for this job
to only run once per mentor. Each time it runs,
the mentees will be reassigned, which notifies
the affected mentees (repeated reassignments can
then be fairly annoying).

It has been reported that ReassignMenteesJob can
run more than once for one mentor. As of today,
we don't know why that happened. Let's prevent
re-occurances via jobqueue-based deduplication
for now. While it doesn't guarantee double reassignment
won't happen in the future, it should significantly decrease
the risk of a similar bug happening again.

Bug: T322374
Bug: T323374
Change-Id: Ic42b9051ac1d6d09d0ebff7ceb85a09631ffbeb8