Makeshift deduplication for RefreshUserImpactJob
Deduplicating job specifiers doesn't work for batch jobs; instead
check if there's a fresh value in the cache before trying to
generate it.

Takes a somewhat lazy approach to handling batch user impact
loading, to avoid having to implement it in all UserImpactLookup
classes. We might want to improve that in a follow-up.

Bug: T322706
Change-Id: I7331005ed6abd0275264cfe1cfb007c3e7d03683