UserImpact: Set a specific end date for pageviews application URL
"latest" can be open to interpretation, and lead to discrepancies
between what the user sees in the impact module table and in the
application. Let's use the latest available page view data date as the
"end" parameter to ensure consistency between the Impact module display
and the link the user follows to the pageviews application.

Bug: T323748
Bug: T323253
Change-Id: Iafef24d3541150c8f8902e92654d102c06c4f568